+++ examples/encrypted-chat/src/hooks/useView.tsx
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+
+export enum VIEWS {
+  Login = "LOGIN",
+  Dashboard = "DASHBOARD"
+}
+
+export const useView = () => {
+  const [view, setView] = useState(VIEWS.Login);
+
+  return {view, setView};
+};
